+++ packages/dds/tree/src/test/feature-libraries/flex-tree/utils.ts
@@ -21,7 +21,7 @@ import {
 } from "../../../feature-libraries/index.js";
 import { TreeContent } from "../../../shared-tree/index.js";
 import { MockTreeCheckout, forestWithContent } from "../../utils.js";
-import { MockNodeKeyManager } from "../../MockNodeKeyManager.js";
+import { MockNodeKeyManager } from "../../nodeKeyManager.js";
 
 export function getReadonlyContext(forest: IEditableForest, schema: FlexTreeSchema): Context {
 	return getTreeContext(schema, new MockTreeCheckout(forest), new MockNodeKeyManager());
