+++ packages/dds/tree/src/test/shared-tree/sharedTree.spec.ts
@@ -91,7 +91,7 @@ import {
 } from "../utils.js";
 import { configuredSharedTree } from "../../treeFactory.js";
 import { ISharedObjectKind } from "@fluidframework/shared-object-base/internal";
-import { MockNodeKeyManager } from "../MockNodeKeyManager.js";
+import { MockNodeKeyManager } from "../nodeKeyManager.js";
 
 const DebugSharedTree = configuredSharedTree({
 	jsonValidator: typeboxValidator,
